pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "good-first-issue"
version = "0.1.0"
description = "Find good first issues in open source projects you actually care about"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Jacob", email = "your.email@example.com" }
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "Pillow>=10.0.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
gfi = "gfi.cli:cli"

[tool.black]
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["gfi"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
